Further migration over to TS
